Avoid double-importing strategy modules by resolver
This case was noticed when driving freqtrade from the user codebase.

This trips up tests as they do things like `freqtrade.config["order_types"]["exit"] = "market"` which modifies a dictionary on `StrategyTestV3`. The problem is that the value is both invalid according to config validation, and results in future tests having state they were not expecting. An example of this issue in play can be seen in `test_execute_trade_exit_market_order` where the first test works, then the second fails due to the one before it - similarly other tests will fail afterward.