Add `config.get("single_load", False)` as switch for single-load behaviour
This seemed to most sane option for this, as opposed to defaulting to be one and patching all the tests to use monkeypatch's setitem or setattr as relevant as the change itself would be major for at least some users.

This also makes it so the module will be reused even if it hasn't been imported by some normal import (see sys.modules assignment).